fix(W2.B refactor): cast embeddings dimensions to int + sync .env.example
Two Copilot round-2 findings on PR #84:

1. **`config/ai.php` `models.embeddings.dimensions`** read
   `env('REGOLO_EMBEDDINGS_DIMENSIONS', 4096)` without a cast.
   `env()` returns a STRING from `.env` files (`"4096"`), not the
   default's int. The SDK's `EmbeddingProvider::defaultEmbeddingsDimensions(): int`
   contract would crash on a strict TypeError when an adopter sets
   the var. Added the explicit `(int)` cast with an inline comment
   explaining why `env()` casts aren't free.

2. **`.env.example` Regolo block was misaligned with the new
   config-key surface.** The previous block listed `REGOLO_TEMPERATURE`
   and `REGOLO_MAX_TOKENS` (both unused by the laravel/ai SDK
   adoption) and was missing `REGOLO_CHAT_MODEL_CHEAPEST`,
   `REGOLO_CHAT_MODEL_SMARTEST`, `REGOLO_EMBEDDINGS_DIMENSIONS`,
   `REGOLO_RERANKING_MODEL`. Updated the section to:

   - List every env var the new `config/ai.php` block reads.
   - Drop the unused temperature / max-tokens vars.
   - Add a footnote explaining the temperature / max-tokens absence
     (the SDK's `AnonymousAgent::prompt()` does not expose those
     setters; per-call options propagation is tracked for
     v4.0 W2.D follow-up).

The R40 test count is unchanged (8 / 21 — the role-guard test added
in commit `1aff275` already covered the related behaviour).

+++ config/ai.php
@@ -96,7 +96,11 @@ return [
                 ],
                 'embeddings' => [
                     'default' => env('REGOLO_EMBEDDINGS_MODEL', 'Qwen3-Embedding-8B'),
-                    'dimensions' => env('REGOLO_EMBEDDINGS_DIMENSIONS', 4096),
+                    // env() returns a string from .env files; cast to int
+                    // because the SDK signature is `int $dimensions` and a
+                    // string value would crash on a strict TypeError when
+                    // the embeddings gateway is invoked.
+                    'dimensions' => (int) env('REGOLO_EMBEDDINGS_DIMENSIONS', 4096),
                 ],
                 'reranking' => [
                     'default' => env('REGOLO_RERANKING_MODEL', 'jina-reranker-v2'),
